+++ mood-garden/src/components/NextPlant/NextPlant.tsx
