+++ mood-garden/src/pages/Progress.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import Calendar from "react-calendar";
-import { NextPlant } from "./components/NextPlant/NextPlant";
+import { NextPlant } from "../components/NextPlant/NextPlant";
 import "react-calendar/dist/Calendar.css";
 import "../../styles/progress.css";
 import { View } from "react-calendar/dist/cjs/shared/types";
